Rename settings tests to SettingsPage and update imports
